# Conflicts:
#	src/main/java/lt/baltictalents/p23/demo1/Demo1.java
#	src/main/java/lt/baltictalents/p23/demo3/Demo3.java
#	src/main/java/lt/baltictalents/p23/demo4/Demo4.java
#	src/main/java/lt/baltictalents/p23/demo5/Demo5.java
#	src/main/java/lt/baltictalents/p23/demo6/Demo6.java

+++ src/main/java/lt/baltictalents/p29/demo2/Salary.java
@@ -5,23 +5,23 @@ import java.time.LocalDate;
 
 public class Salary {
 
-    private double amount;
+    private BigDecimal amount;
 
     private LocalDate hired;
 
     private LocalDate fired;
 
-    public Salary(double amount, LocalDate hired, LocalDate fired) {
+    public Salary(BigDecimal amount, LocalDate hired, LocalDate fired) {
         this.amount = amount;
         this.hired = hired;
         this.fired = fired;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
